Fix an issue in download requests where `urlPath` is not getting initialized properly
Summary:
This is an attempt to fix the following issues:
 - https://github.com/dropbox/SwiftyDropbox/issues/30
 - https://github.com/dropbox/SwiftyDropbox/issues/31

A best-guess (cannot repro this issue locally, seems to be a timing issue?) at what has occurred is that when the download request is constructed, it starts, and in the worst case this potentially means that the reference to `self` is not initialized properly before the completion closure is called. This means we lose the `self.urlPath`, causing unwrapping issues.

This attempts to fix the problem by only starting the request after we have finished initializing the reference to `self`.

Reviewers: wkiefer

Reviewed By: wkiefer

Differential Revision: https://tails.corp.dropbox.com/D154526