[SwiftyDropbox] Refactored BabelClient
Summary:
As part of implementing team auth for SwiftyDropbox and open-sourcing Babel, this diff changes the way we implement our networking client. Instead of having a "BabelClient" that lives in the Babel repo, we move the BabelClient (Client.swift) to the SwiftyDropbox repo and rename it to DropboxTransportClient (DropboxTransportClient.swift). This accomplishes a number of things.

- it dispels the notion that the BabelClient base is somehow specific to Babel and not significantly dependent on assumptions about Dropbox API
- allows us to get rid of the wrapper DropboxClient layer (and in the Babel repo, the DropboxClient-raw.swift file)

This diff also separates the auto-generated code for the Dropbox (user) client into two files: DropboxClientCommon.swift (which is the auto-generated code and serves as the parent of DropboxClient) and DropboxClient.swift (which has the handwritten code and serves as the child of DropboxClientCommon). This design paradigm makes it clearer which code is generated and which is not (and avoids having generator code that is emiting raw, hardcoded unstructured code)

Corresponding Babel repo diff here: D199513

Test Plan: ran unit tests

Reviewers: leahculver, greg, kelkabany, jlee

Reviewed By: jlee

Differential Revision: https://tails.corp.dropbox.com/D199509