[SwiftyDropbox] [SwiftyDropbox repo] Harmonize SwiftyDropbox with Stone
Summary:
This diff harmonizes SwiftyDropbox with Stone. Namely, all stone code generation is done in this repository now. Both stone repos and stone spec repos are now submodules of this repo.

In addition, the Dropbox layer is more thoroughly separated from the Stone layer:

- Base.swift and BaseTeam.swift (which are auto-generated classes that contain the API routes) are entirely separated from DropboxClient.swift and DropboxTeamClient.swift (which have handwritten client logic inside of them)
- DropboxTransportClient and additional client-side route arguments are now passed into the generator as command line arguments. These client-side arguments are then injected into routes based on route style. This is done because stone specs are unable to entirely anticipate client side needs
- Client-side arguments are wrapped in an auto-generated ClientArgument wrappers, before being passed to DropboxTransportClient

The vast majority of this diff is implementing stone spec updates.

This diff goes in combination with the Stone repo diff: D202297.

Test Plan: Unit tests

Reviewers: kelkabany, jlee

Reviewed By: jlee

Differential Revision: https://tails.corp.dropbox.com/D202309