Added support to show loading status for the last step of OAuth code flow
OAuth code flow requires a network request to get access token, apps should build some UX to handle this async operation.
A default UX is built in the SDK with showing a load spinner while waiting for oauth token request to finish.
Apps can provide a `LoadingStatusDelegate` to override the default and show a custom experience.